+++ lib/rubocop/cop/style/rescue_modifier.rb
@@ -10,7 +10,8 @@ module RuboCop
         MSG = 'Avoid using `rescue` in its modifier form.'
 
         def investigate(processed_source)
-          @modifier_locations = processed_source.tokens
+          @modifier_locations = processed_source
+                                .tokens
                                 .select { |t| t.type == :kRESCUE_MOD }
                                 .map(&:pos)
         end