+++ spec/rubocop/config_store_spec.rb
@@ -30,13 +30,13 @@ describe RuboCop::ConfigStore do
 
     context 'when no config specified in command line' do
       it 'gets config path and config from cache if available' do
-        expect(RuboCop::ConfigLoader).to receive(:configuration_file_for).once
-          .with('dir')
-        expect(RuboCop::ConfigLoader).to receive(:configuration_file_for).once
-          .with('dir/subdir')
+        expect(RuboCop::ConfigLoader)
+          .to receive(:configuration_file_for).once.with('dir')
+        expect(RuboCop::ConfigLoader)
+          .to receive(:configuration_file_for).once.with('dir/subdir')
         # The stub returns the same config path for dir and dir/subdir.
-        expect(RuboCop::ConfigLoader).to receive(:configuration_from_file).once
-          .with('dir/.rubocop.yml')
+        expect(RuboCop::ConfigLoader)
+          .to receive(:configuration_from_file).once.with('dir/.rubocop.yml')
 
         config_store.for('dir/file2')
         config_store.for('dir/file2')