+++ spec/rubocop/cop/style/multiline_operation_indentation_spec.rb
@@ -33,20 +33,30 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'registers an offense for one space indentation of second line' do
       inspect_source(cop,
-                     ['a',
-                      ' .b'])
+                     ['a +',
+                      ' b'])
       expect(cop.messages).to eq(['Use 2 (not 1) spaces for indenting an ' \
                                   'expression spanning multiple lines.'])
-      expect(cop.highlights).to eq(['.b'])
+      expect(cop.highlights).to eq(['b'])
     end
 
-    it 'registers an offense for proc call without a selector' do
+    it 'does not check method calls' do
       inspect_source(cop,
                      ['a',
-                      ' .(args)'])
-      expect(cop.messages).to eq(['Use 2 (not 1) spaces for indenting an ' \
-                                  'expression spanning multiple lines.'])
-      expect(cop.highlights).to eq(['.('])
+                      ' .(args)',
+                      '',
+                      'Foo',
+                      '.a',
+                      '  .b',
+                      '',
+                      'Foo',
+                      '.a',
+                      '  .b(c)',
+                      '',
+                      'expect { Foo.new }.',
+                      '  to change { Bar.count }.',
+                      '      from(1).to(2)'])
+      expect(cop.offenses).to be_empty
     end
 
     it 'registers an offense for three spaces indentation of second line' do
@@ -80,6 +90,16 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
       expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
                                   'expression spanning multiple lines.'])
       expect(cop.highlights).to eq(['b'])
+    end
+
+    it 'accepts indented operands in an array' do
+      inspect_source(cop, ['    dm[i][j] = [',
+                           '      dm[i-1][j-1] +',
+                           '        (this[j-1] == that[i-1] ? 0 : sub),',
+                           '      dm[i][j-1] + ins,',
+                           '      dm[i-1][j] + del',
+                           '    ].min'])
+      expect(cop.offenses).to be_empty
     end
 
     it 'accepts two spaces indentation of second line' do
@@ -121,36 +141,6 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
       expect(cop.messages).to be_empty
     end
 
-    it 'registers an offense for badly indented operands in chained ' \
-       'method call' do
-      inspect_source(cop,
-                     ['Foo',
-                      '.a',
-                      '  .b'])
-      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
-                                  'expression spanning multiple lines.'])
-      expect(cop.highlights).to eq(['.a'])
-    end
-
-    it 'registers an offense for badly indented operands in chained ' \
-       'method call' do
-      inspect_source(cop,
-                     ['Foo',
-                      '.a',
-                      '  .b(c)'])
-      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
-                                  'expression spanning multiple lines.'])
-      expect(cop.highlights).to eq(['.a'])
-    end
-
-    it 'accepts even indentation of consecutive lines in typical RSpec code' do
-      inspect_source(cop,
-                     ['expect { Foo.new }.',
-                      '  to change { Bar.count }.',
-                      '  from(1).to(2)'])
-      expect(cop.messages).to be_empty
-    end
-
     it 'accepts any indentation of parameters to #[]' do
       inspect_source(cop,
                      ['payment = Models::IncomingPayments[',
@@ -159,31 +149,13 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
       expect(cop.messages).to be_empty
     end
 
-    it 'registers an offense for extra indentation of 3rd line in typical ' \
-       'RSpec code' do
-      inspect_source(cop,
-                     ['expect { Foo.new }.',
-                      '  to change { Bar.count }.',
-                      '      from(1).to(2)'])
-      expect(cop.messages).to eq(['Use 2 (not 6) spaces for indenting an ' \
+    it 'registers an offense for an unindented multiline operation that is ' \
+       'the left operand in another operation' do
+      inspect_source(cop, ['a +',
+                           'b < 3'])
+      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
                                   'expression spanning multiple lines.'])
-    end
-
-    it 'registers an offense for aligned code on LHS of assignment' do
-      inspect_source(cop, ['def config_to_allow_offenses',
-                           '  Formatter::DisabledConfigFormatter',
-                           '  .config_to_allow_offenses[cop_name] ||= {}',
-                           'end'])
-      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
-        'expression spanning multiple lines.'])
-    end
-
-    it 'registers an offense for a method call where the receiver and ' \
-       'selector are on different lines, without indentation' do
-      inspect_source(cop, ['Formatter::DisabledConfigFormatter',
-                           '.config_to_allow_offenses[cop_name] ||= {}'])
-      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
-        'expression spanning multiple lines.'])
+      expect(cop.highlights).to eq(['b'])
     end
   end
 
@@ -201,7 +173,7 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
       expect(cop.messages).to be_empty
     end
 
-    it 'registers an offense for misaligned operands in if condition' do
+    it 'registers an offense for indented operands in if condition' do
       inspect_source(cop,
                      ['if a +',
                       '    b',
@@ -213,12 +185,57 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
       expect(cop.config_to_allow_offenses).to eq('EnforcedStyle' => 'indented')
     end
 
+    it 'accepts indented code on LHS of equality operator' do
+      inspect_source(cop, ['def config_to_allow_offenses',
+                           '  a +',
+                           '    b == c ',
+                           'end'])
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'accepts indented operands inside block + assignment' do
+      inspect_source(cop, ['a = b.map do |c|',
+                           '  c +',
+                           '    d',
+                           'end',
+                           '',
+                           'requires_interpolation = node.children.any? do |s|',
+                           '  s.type == :dstr ||',
+                           '    s.loc.expression.source =~ REGEXP',
+                           'end'])
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'registers an offense for indented second part of string' do
+      inspect_source(cop,
+                     ['it "should convert " +',
+                      '  "a to " +',
+                      '  "b" do',
+                      'end'])
+      expect(cop.messages).to eq(['Align the operands of an expression ' \
+                                  'spanning multiple lines.'] * 2)
+      expect(cop.highlights).to eq(['"a to "', '"b"'])
+      expect(cop.config_to_allow_offenses).to eq('EnforcedStyle' => 'indented')
+    end
+
+    it 'registers an offense for indented operand in second argument' do
+      inspect_source(cop,
+                     ['puts a, 1 +',
+                      '  2'])
+      expect(cop.messages)
+        .to eq(['Align the operands of an expression spanning multiple lines.'])
+      expect(cop.highlights).to eq(['2'])
+      expect(cop.config_to_allow_offenses).to eq('EnforcedStyle' => 'indented')
+    end
+
     it 'registers an offense for misaligned string operand when the first ' \
        'operand has backslash continuation' do
       inspect_source(cop,
-                     ["flash[:error] = 'Here is a string ' \\",
-                      "                'That spans' <<",
-                      "  'multiple lines'"])
+                     ['def f',
+                      "  flash[:error] = 'Here is a string ' \\",
+                      "                  'That spans' <<",
+                      "    'multiple lines'",
+                      'end'])
       expect(cop.messages).to eq(['Align the operands of an expression in an ' \
                                   'assignment spanning multiple lines.'])
       expect(cop.highlights).to eq(["'multiple lines'"])
@@ -226,8 +243,8 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'registers an offense for misaligned string operand when plus is used' do
       inspect_source(cop,
-                     ["flash[:error] = 'Here is a string ' +",
-                      "                'That spans' <<",
+                     ["Error = 'Here is a string ' +",
+                      "        'That spans' <<",
                       "  'multiple lines'"])
       expect(cop.messages).to eq(['Align the operands of an expression in an ' \
                                   'assignment spanning multiple lines.'])
@@ -236,64 +253,70 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'registers an offense for misaligned operands in unless condition' do
       inspect_source(cop,
-                     ['unless a',
-                      '  .b',
+                     ['unless a +',
+                      '  b',
                       '  something',
                       'end'])
       expect(cop.messages).to eq(['Align the operands of a condition in an ' \
                                   '`unless` statement spanning multiple ' \
                                   'lines.'])
-      expect(cop.highlights).to eq(['.b'])
+      expect(cop.highlights).to eq(['b'])
       expect(cop.config_to_allow_offenses).to eq('Enabled' => false)
     end
 
-    it 'registers an offense for misaligned operands in while condition' do
-      inspect_source(cop,
-                     ['while a.',
-                      '    b',
-                      '  something',
-                      'end'])
-      expect(cop.messages).to eq(['Align the operands of a condition in a ' \
-                                  '`while` statement spanning multiple lines.'])
-      expect(cop.highlights).to eq(['b'])
-    end
-
-    it 'registers an offense for misaligned operands in until condition' do
-      inspect_source(cop,
-                     ['until a.',
-                      '    b',
-                      '  something',
-                      'end'])
-      expect(cop.messages).to eq(['Align the operands of a condition in an ' \
-                                  '`until` statement spanning multiple lines.'])
-      expect(cop.highlights).to eq(['b'])
+    [
+      %w(an if),
+      %w(an unless),
+      %w(a while),
+      %w(an until)
+    ].each do |article, keyword|
+      it "registers an offense for misaligned operands in #{keyword} " \
+         'condition' do
+        inspect_source(cop,
+                       ["#{keyword} a or",
+                        '    b',
+                        '  something',
+                        'end'])
+        expect(cop.messages).to eq(['Align the operands of a condition in ' \
+                                    "#{article} `#{keyword}` statement " \
+                                    'spanning multiple lines.'])
+        expect(cop.highlights).to eq(['b'])
+        expect(cop.config_to_allow_offenses)
+          .to eq('EnforcedStyle' => 'indented')
+      end
     end
 
     it 'accepts aligned operands in assignment' do
       inspect_source(cop,
-                     ['formatted_int = int_part',
-                      '                .to_s',
-                      '                .reverse',
-                      "                .gsub(/...(?=.)/, '\&_')"])
+                     ['a = b +',
+                      '    c +',
+                      '    d'])
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'accepts aligned or:ed operands in assignment' do
+      inspect_source(cop,
+                     ["tmp_dir = ENV['TMPDIR'] || ENV['TMP'] || ENV['TEMP'] ||",
+                      "          Etc.systmpdir || '/tmp'"])
       expect(cop.messages).to be_empty
     end
 
-    it 'registers an offense for unaligned operands in assignment' do
+    it 'registers an offense for unaligned operands in op-assignment' do
       inspect_source(cop,
-                     ['bar = Foo',
-                      '  .a',
-                      '      .b(c)'])
+                     ['bar *= Foo +',
+                      '  a +',
+                      '       b(c)'])
       expect(cop.messages).to eq(['Align the operands of an expression in an ' \
                                   'assignment spanning multiple lines.'])
-      expect(cop.highlights).to eq(['.a'])
+      expect(cop.highlights).to eq(['a'])
     end
 
     it 'auto-corrects' do
-      new_source = autocorrect_source(cop, ['until a.',
+      new_source = autocorrect_source(cop, ['until a +',
                                             '    b',
                                             '  something',
                                             'end'])
-      expect(new_source).to eq(['until a.',
+      expect(new_source).to eq(['until a +',
                                 '      b',
                                 '  something',
                                 'end'].join("\n"))
@@ -344,9 +367,9 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'accepts any indentation of method parameters' do
       inspect_source(cop,
-                     ['a(b.',
-                      '    c',
-                      '.d)'])
+                     ['a(b +',
+                      '    c +',
+                      'd)'])
       expect(cop.messages).to be_empty
     end
 
@@ -356,7 +379,16 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
                       '  arg_array == a ||',
                       '  arg_array.map(&:children) == a.map(&:children)',
                       ')'])
-      expect(cop.messages).to be_empty
+      expect(cop.offenses).to be_empty
+    end
+
+    it 'registers an offense for aligned code on LHS of equality operator' do
+      inspect_source(cop, ['def config_to_allow_offenses',
+                           '  a +',
+                           '  b == c ',
+                           'end'])
+      expect(cop.messages).to eq(['Use 2 (not 0) spaces for indenting an ' \
+                                  'expression spanning multiple lines.'])
     end
 
     [
@@ -383,16 +415,12 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
                        ["#{keyword} receiver.nil? &&",
                         '  !args.empty? &&',
                         '  BLACKLIST.include?(method_name)',
-                        'end',
-                        "#{keyword} receiver.",
-                        '  nil?',
                         'end'])
         expect(cop.highlights).to eq(['!args.empty?',
-                                      'BLACKLIST.include?(method_name)',
-                                      'nil?'])
+                                      'BLACKLIST.include?(method_name)'])
         expect(cop.messages).to eq(['Use 4 (not 2) spaces for indenting a ' \
                                     "condition in #{article} `#{keyword}` " \
-                                    'statement spanning multiple lines.'] * 3)
+                                    'statement spanning multiple lines.'] * 2)
       end
 
       it "accepts indented operands in #{keyword} body" do
@@ -436,12 +464,9 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'accepts indentation of assignment' do
       inspect_source(cop,
-                     ['formatted_int = int_part',
-                      '  .abs',
-                      '  .to_s',
-                      '  .reverse',
-                      "  .gsub(/...(?=.)/, '\&_')",
-                      '  .reverse'])
+                     ['a = b +',
+                      '  c +',
+                      '  d'])
       expect(cop.messages).to be_empty
     end
 
@@ -459,20 +484,20 @@ describe RuboCop::Cop::Style::MultilineOperationIndentation do
 
     it 'registers an offense for aligned operators in assignment' do
       inspect_source(cop,
-                     ['formatted_int = int_part',
-                      '                .abs',
-                      '                .reverse'])
-      expect(cop.messages).to eq(['Use 2 (not 16) spaces for indenting an ' \
+                     ['a = b +',
+                      '    c +',
+                      '    d'])
+      expect(cop.messages).to eq(['Use 2 (not 4) spaces for indenting an ' \
                                   'expression in an assignment spanning ' \
                                   'multiple lines.'] * 2)
     end
 
     it 'auto-corrects' do
-      new_source = autocorrect_source(cop, ['until a.',
+      new_source = autocorrect_source(cop, ['until a +',
                                             '      b',
                                             '  something',
                                             'end'])
-      expect(new_source).to eq(['until a.',
+      expect(new_source).to eq(['until a +',
                                 '    b',
                                 '  something',
                                 'end'].join("\n"))